Combine histogram page and table page
