--- src/web/app/pages-logs/logs-histogram-page/logs-histogram-page.component.html
@@ -1,56 +1,0 @@
-<h1>Logs Histogram View</h1>
-<p>This page provides a histogram view for error logs in the past {{ LOGS_RETENTION_PERIOD_IN_DAYS }} days.</p>
-<button (click)="toLogsPage()" class="btn btn-info">Go to logs page</button>
-<hr/>
-
-<div *tmIsLoading="isLoading" class="card bg-form no-border">
-  <div class="card-body">
-    <div class="row form-group">
-      <div class="col-5">
-        <div class="col-12 mb-1">
-          <label for="logs-from-datepicker" class="font-weight-bold">Search period from</label>
-        </div>
-        <div class="input-group">
-          <div class="col-md-6 input-group">
-            <input id="logs-from-datepicker" type="text" class="form-control" ngbDatepicker [minDate]="earliestSearchDate" [maxDate]="dateToday" [(ngModel)]="formModel.logsDateFrom" #logsFromDp="ngbDatepicker"/>
-            <button class="btn btn-light" (click)="logsFromDp.toggle()" type="button">
-              <i class="fas fa-calendar-alt"></i>
-            </button>
-          </div>
-          <div id="logs-from-timepicker" class="col-md-4">
-            <tm-time-picker [(time)]="formModel.logsTimeFrom"></tm-time-picker>
-          </div>
-        </div>
-      </div>
-      <div class="col-5">
-        <div class="col-12 mb-1">
-          <label for="logs-to-datepicker" class="font-weight-bold">Search period until</label>
-        </div>
-        <div class="input-group">
-          <div class="col-md-6 input-group">
-            <input id="logs-to-datepicker" type="text" class="form-control" ngbDatepicker [minDate]="earliestSearchDate" [maxDate]="dateToday" [(ngModel)]="formModel.logsDateTo" #logsToDp="ngbDatepicker"/>
-            <button class="btn btn-light" (click)="logsToDp.toggle()" type="button">
-              <i class="fas fa-calendar-alt"></i>
-            </button>
-          </div>
-          <div id="logs-to-timepicker" class="col-md-4">
-            <tm-time-picker [(time)]="formModel.logsTimeTo"></tm-time-picker>
-          </div>
-        </div>
-      </div>
-      <div class="col-1 mt-4">
-        <button id="query-button" class="float-right btn btn-primary" (click)="searchForLogs()">Query</button>
-      </div>
-    </div>
-  </div>
-</div>
-
-<div id="histogram" style="height: 400px; width: 100%;" *tmIsLoading="isSearching">
-  <tm-logs-histogram [data]="searchResult" *ngIf="searchResult.length>0; else noDataMessage"></tm-logs-histogram>
-</div>
-
-<ng-template #noDataMessage>
-  <div class="no-data-message">
-    There are no error logs in this time period.
-  </div>
-</ng-template>
