--- src/web/app/pages-logs/logs-histogram-page/logs-histogram-page.component.scss
@@ -1,12 +1,0 @@
-.no-border {
-  border: 0;
-}
-
-.bg-form {
-  background-color: #EAEFF5;
-}
-
-.no-data-message {
-  padding: 10px;
-  text-align: center;
-}
