+++ src/web/app/pages-logs/logs-page.component.scss
@@ -9,3 +9,8 @@
 .center {
   text-align: center;
 }
+
+.no-data-message {
+  padding: 10px;
+  text-align: center;
+}
