+++ src/web/app/pages-logs/logs-page.component.spec.ts
@@ -99,7 +99,7 @@ describe('LogsPageComponent', () => {
     expect(logSpy).toHaveBeenCalledWith({
       searchFrom: '0',
       searchUntil: '0',
-      severities: 'INFO',
+      severity: 'INFO',
     });
 
     expect(component.searchResults.length).toEqual(1);
@@ -112,5 +112,82 @@ describe('LogsPageComponent', () => {
     fixture.detectChanges();
     const button: any = fixture.debugElement.nativeElement.querySelector('#load-button');
     expect(button.disabled).toBeTruthy();
+  });
+
+  it('should search for all error logs when search button is clicked', () => {
+    const logSpy: Spy = spyOn(logService, 'searchLogs').and
+        .returnValues(of({ logEntries: [], nextPageToken: 'token' }), of({ logEntries: [] }));
+    const timeSpy: Spy = spyOn(timezoneService, 'getResolvedTimestamp').and
+        .returnValue(of({ timestamp: 0, message: '' }));
+
+    component.isLoading = false;
+    component.isSearching = false;
+    component.isTableView = false;
+    component.formModel = {
+      logsSeverity: new Set(),
+      logsDateFrom: { year: 2021, month: 6, day: 1 },
+      logsTimeFrom: { hour: 23, minute: 59 },
+      logsDateTo: { year: 2021, month: 6, day: 2 },
+      logsTimeTo: { hour: 23, minute: 59 },
+    };
+    fixture.detectChanges();
+
+    fixture.debugElement.nativeElement.querySelector('#query-button').click();
+
+    expect(timeSpy).toHaveBeenCalledTimes(2);
+    expect(logSpy).toHaveBeenCalledTimes(2);
+    expect(logSpy.calls.mostRecent().args).toEqual([{
+      searchFrom: '0',
+      searchUntil: '0',
+      severity: 'ERROR',
+      nextPageToken: 'token',
+    }]);
+  });
+
+  it('should sort logs based on source location', () => {
+    component.isTableView = false;
+    const testLog1: GeneralLogEntry = {
+      logName: 'stderr',
+      severity: 'ERROR',
+      trace: 'testTrace1',
+      sourceLocation: {
+        file: 'file1',
+        line: 10,
+        function: 'function1',
+      },
+      timestamp: 1549095330000,
+      message: 'message',
+    };
+    const testLog2: GeneralLogEntry = {
+      logName: 'stderr',
+      severity: 'ERROR',
+      trace: 'testTrace2',
+      sourceLocation: {
+        file: 'file2',
+        line: 10,
+        function: 'function1',
+      },
+      timestamp: 1549095330000,
+      message: 'message',
+    };
+    const testLog3: GeneralLogEntry = {
+      logName: 'stderr',
+      severity: 'ERROR',
+      trace: 'testTrace3',
+      sourceLocation: {
+        file: 'file2',
+        line: 10,
+        function: 'function1',
+      },
+      timestamp: 1549095330000,
+      message: 'message',
+    };
+    spyOn(logService, 'searchLogs').and
+      .returnValue(of({ logEntries: [testLog1, testLog2, testLog3] }));
+    spyOn(timezoneService, 'getResolvedTimestamp').and.returnValue(of({ timestamp: 0, message: '' }));
+
+    component.searchForLogs();
+
+    expect(component.histogramResult.length).toEqual(2);
   });
 });
