+++ src/web/app/pages-logs/logs-page.module.ts
@@ -4,6 +4,7 @@ import { FormsModule } from '@angular/forms';
 import { RouterModule, Routes } from '@angular/router';
 import { NgbDatepickerModule } from '@ng-bootstrap/ng-bootstrap';
 import { LoadingSpinnerModule } from '../components/loading-spinner/loading-spinner.module';
+import { LogsHistogramModule } from '../components/logs-histogram/logs-histogram.module';
 import { LogsTableModule } from '../components/logs-table/logs-table.module';
 import { SessionEditFormModule } from '../components/session-edit-form/session-edit-form.module';
 import { SortableTableModule } from '../components/sortable-table/sortable-table.module';
@@ -13,11 +14,6 @@ const routes: Routes = [
   {
     path: '',
     component: LogsPageComponent,
-  },
-  {
-    path: 'histogram',
-    loadChildren: () => import('./logs-histogram-page/logs-histogram-page.module')
-        .then((m: any) => m.LogsHistogramPageModule),
   },
 ];
 
@@ -34,6 +30,7 @@ const routes: Routes = [
     LogsTableModule,
     NgbDatepickerModule,
     SessionEditFormModule,
+    LogsHistogramModule,
   ],
   declarations: [
     LogsPageComponent,
